% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal_Hist2D.R
\name{addRightMargPlot}
\alias{addRightMargPlot}
\title{add a marginal plot to the right side of another plot}
\usage{
addRightMargPlot(ggMargGrob, right, size)
}
\arguments{
\item{ggMargGrob}{- plot grob to add marginal plot to}

\item{right}{- marginal plot to add}

\item{size}{- relative width of main plot to marginal plot}
}
\description{
Function to add a marginal plot to the right of another plot.
}
\details{
Copied from \code{ggExtra}. Created by Dean Attali ("daattali@gmail.com").
}
\keyword{internal}
