% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal_Hist2D.R
\name{addTitleGrobs}
\alias{addTitleGrobs}
\title{Add the title/subtitle grobs to the main plot grob, along with a little padding}
\usage{
addTitleGrobs(gtbl, titleGrobs)
}
\arguments{
\item{gtbl}{- gtable to add title grobs to}

\item{titleGrobs}{- grobs to add to gtable}

\item{l}{- left size}

\item{r}{- right size}
}
\description{
Function to add the title/subtitle grobs to the main plot grob, 
along with a little padding.
}
\details{
Modified from \code{ggExtra}. Created by Dean Attali ("daattali@gmail.com").
}
\keyword{internal}
