% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal_Hist2D.R
\name{addTopMargPlot}
\alias{addTopMargPlot}
\title{add a marginal plot to the top of another plot}
\usage{
addTopMargPlot(ggMargGrob, top, size)
}
\arguments{
\item{ggMargGrob}{- plot grob to add marginal plot to}

\item{top}{- marginal plot to add}

\item{size}{- relative height of main plot to marginal plot}
}
\description{
Function to add a marginal plot to the top of another plot.
}
\details{
Copied from \code{ggExtra}. Created by Dean Attali ("daattali@gmail.com").
}
\keyword{internal}
