% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_functions.R
\name{getLegend}
\alias{getLegend}
\title{Get the plot legend(s)}
\usage{
getLegend(plot, return_all = FALSE)
}
\arguments{
\item{plot}{- object capable of being converted to a gtable (e.g., a ggplot2 object)}

\item{return_all}{- flag to return all legends}
}
\value{
a legend grob or list of grobs.
}
\description{
Function to get the plot legend(s)  (based on cowplot:::get_legend)
}
\details{
Port of cowplot:::get_legend
}
