% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal_Hist2D.R
\name{getMargGrob}
\alias{getMargGrob}
\title{Get the panel grob from a ggplot2 plot object}
\usage{
getMargGrob(plt)
}
\arguments{
\item{plt}{- ggplot2 plot object}
}
\value{
a gtable with the panel grob only.
}
\description{
Function to get the panel grob from a ggplot2 plot object.
}
\details{
Modified from \code{ggExtra}. Created by Dean Attali ("daattali@gmail.com").
}
\keyword{internal}
