% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal_Hist2D.R
\name{getPanelPos}
\alias{getPanelPos}
\title{Get the panel position}
\usage{
getPanelPos(gtbl)
}
\arguments{
\item{gtbl}{- a gtable}
}
\value{
Position of panel (t,l,b,r).
}
\description{
Function to add a grob to the top of a gtable.
}
\details{
Copied from \code{ggExtra}. Created by Dean Attali ("daattali@gmail.com").
}
\keyword{internal}
