% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_functions.R
\name{getPlotComponent}
\alias{getPlotComponent}
\title{Get a plot component}
\usage{
getPlotComponent(plot, pattern, return_all = FALSE)
}
\arguments{
\item{plot}{- object capable of being converted to a gtable (e.g., a ggplot2 object)}

\item{pattern}{- grob pattern to look for}

\item{return_all}{-  flag to return all objects matching the pattern, otherwise just the first}
}
\value{
invisibly, a grob or list of grobs matching the pattern.
}
\description{
Function to get a plot component (based on cowplot:::get_plot_compnent)
}
\details{
Port of cowplot:::get_plot_component
}
