% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStdTheme.R
\name{getStdTheme}
\alias{getStdTheme}
\title{Get standard plot theme}
\usage{
getStdTheme()
}
\value{
\pkg{ggplot2} theme object with "standard" thematics
}
\description{
Function to get the standard plot theme
}
