% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal_Hist2D.R
\name{getTitleGrobs}
\alias{getTitleGrobs}
\title{Pull out the title and subtitle grobs for a ggplot2 plot}
\usage{
getTitleGrobs(p)
}
\arguments{
\item{p}{- plot}
}
\description{
Function to pull out the title and subtitle grobs for a ggplot2 plot, after check to
make sure there is a title.
}
\details{
Copied from \code{ggExtra}. Created by Dean Attali ("daattali@gmail.com").
}
\keyword{internal}
