% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_functions.R
\name{noL}
\alias{noL}
\title{Turn off ggplot2 legend}
\usage{
noL()
}
\value{
invisibly, a `ggplot2` theme obj
}
\description{
Function to turn off y-axis title
}
\details{
Shortcut to "ggplot2::theme(legend.position="none")".
}
