% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_functions.R
\name{noXT}
\alias{noXT}
\title{Turn off ggplot2 x-axis title}
\usage{
noXT()
}
\value{
invisibly, a `ggplot2` theme obj
}
\description{
Function to turn off x-axis title
}
\details{
Shortcut to "ggplot2::theme(axis.title.x=element_blank())".
}
