% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_functions.R
\name{noYT}
\alias{noYT}
\title{Turn off ggplot2 y-axis title}
\usage{
noYT()
}
\value{
invisibly, a [ggplot2::theme] object
}
\description{
Function to turn off y-axis title
}
\details{
Shortcut to "ggplot2::theme(axis.title.y=element_blank())".
}
