% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_functions.R
\name{plotComponentNames}
\alias{plotComponentNames}
\title{Get names of all plot components}
\usage{
plotComponentNames(plot)
}
\arguments{
\item{plot}{- object capable of being converted to a gtable (e.g., a ggplot2 object)}
}
\value{
a vector of grob names.
}
\description{
Function to get names of all plot components (based on cowplot:::plot_component_names)
}
\details{
Port of cowplot:::plot_component_names
}
