% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_info_functions.R
\name{plotComponents}
\alias{plotComponents}
\title{Get all plot components}
\usage{
plotComponents(plot)
}
\arguments{
\item{plot}{- object capable of being converted to a gtable (e.g., a ggplot2 object)}
}
\value{
a grob or list of grobs.
}
\description{
Function to get all plot components (based on cowplot:::get_plot_compnent)
}
\details{
Port of cowplot:::plot_components
}
