% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDHARMaResiduals.R
\name{plotDHARMa}
\alias{plotDHARMa}
\title{Create a combined [DHARMa] residuals and QQ plot as a [ggplot2] object}
\usage{
plotDHARMa(
  obj,
  colors = NULL,
  showLegend = FALSE,
  quantreg = NULL,
  rank = TRUE,
  asFactor = NULL,
  smoothScatter = NULL,
  quantiles = c(0.25, 0.5, 0.75),
  shape = 23,
  landscape = FALSE
)
}
\arguments{
\item{obj}{- an object of class "DHARMa"}

\item{colors}{- vector of colors}

\item{showLegend}{- flag to show color/fill legend}

\item{quantreg}{-}

\item{rank}{-}

\item{asFactor}{-}

\item{smoothScatter}{-}

\item{quantiles}{- vector of quantiles to plot as regression lines}

\item{shape}{- [ggplot2] shape code for points}

\item{landscape}{- flag to orient plots in landscape format (default=FALSE)}
}
\value{
list with [ggplot2] plot object and uniformity, dispersion, and outliers tests results
}
\description{
Function to create a combined [DHARMa] residuals and QQ plot as a [ggplot2] object.
}
\details{
Inspired by the R graphics returned by [DHARMa::plotDHARMa], but returned as a ggplot2 plot.
}
