% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDHARMaResiduals.R
\name{plotDHARMaQQ}
\alias{plotDHARMaQQ}
\title{Create a [DHARMa] QQ plot as a [ggplot2] object}
\usage{
plotDHARMaQQ(obj, colors = NULL, showLegend = FALSE, shape = 23)
}
\arguments{
\item{obj}{- an object of class 'DHARMa'}

\item{colors}{- vector of same length as the scaled residuals to use for color-scaling (or NULL)}

\item{showLegend}{- flag to show color/fill legend}

\item{shape}{- ggplot2 shape code (default = 23)}
}
\value{
list with [ggplot2] plot object and uniformity, dispersion, and outliers tests results
}
\description{
Function to create a [DHARMa] QQ plot as a [ggplot2] object.
}
\details{
Inspired by the R graphics returned by [DHARMa::plotQQunif], but returned as a ggplot2 plot.
}
