% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDHARMaResiduals.R
\name{plotDHARMaResiduals}
\alias{plotDHARMaResiduals}
\title{Create a [DHARMa] residuals plot as a [ggplot2] object}
\usage{
plotDHARMaResiduals(
  obj,
  colors = NULL,
  showLegend = FALSE,
  quantreg = NULL,
  rank = TRUE,
  asFactor = NULL,
  smoothScatter = NULL,
  quantiles = c(0.25, 0.5, 0.75)
)
}
\arguments{
\item{obj}{- an object of class 'DHARMa'}

\item{colors}{- vector of colors}

\item{showLegend}{- flag to show color/fill legend}

\item{quantreg}{-}

\item{rank}{-}

\item{asFactor}{-}

\item{smoothScatter}{-}

\item{quantiles}{- vector of quantiles to plot as regression lines}
}
\value{
[ggplot2] plot object
}
\description{
Function to create a [DHARMa] residuals plot as a [ggplot2] object.
}
\details{
Inspired by the R graphics returned by [DHARMa::plotResiduals], but returned as a ggplot2 plot.
}
