% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMC_Diagnostics.R
\name{plotMCMC_Diagnostics}
\alias{plotMCMC_Diagnostics}
\title{Make diagnostic plots for a parameter in an \href{https://mc-stan.org/}{rstan} mcmc simulation}
\usage{
plotMCMC_Diagnostics(mcmc, var, label = var)
}
\arguments{
\item{mcmc}{\itemize{
\item mcmc object compatible with \link{basyesplot} (i.e., from running \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} or \code{\link[tmbstan:tmbstan]{tmbstan::tmbstan()}})
}}

\item{var}{\itemize{
\item name of variable to plot
}}

\item{label}{\itemize{
\item label to use for variable (default=\code{var})
}}
}
\value{
a ggplot2::gg object (or NULL if the \href{https://mc-stan.org/bayesplot/index.html}{bayesplot} package is not installed.)
}
\description{
Function to make diagnostic plots for a parameter in an \href{https://mc-stan.org/}{rstan} mcmc simulation.
}
\details{
Requires the \href{https://mc-stan.org/bayesplot/index.html}{bayesplot} and \link{cowplot} packages.
Combines plots from running \code{\link[bayesplot:MCMC-traces]{bayesplot::mcmc_trace()}}, \code{\link[bayesplot:MCMC-diagnostics]{bayesplot::mcmc_acf_bar()}},
\code{\link[bayesplot:MCMC-distributions]{bayesplot::mcmc_hist()}}, and \code{\link[bayesplot:MCMC-intervals]{bayesplot::mcmc_areas()}} for the model parameter identified
by \code{var} Plots are combined using \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}}. The \code{label} input
can be specified to provide a more informative name for the plotted parameter.
}
