% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.Bubbles.R
\name{plotMDFR.Bubbles}
\alias{plotMDFR.Bubbles}
\title{Plot melted dataframe using a bubble plot format}
\usage{
plotMDFR.Bubbles(
  mdfr,
  x = NULL,
  y = NULL,
  value.var = "val",
  agg.formula = NULL,
  agg.function = sum,
  ...,
  colour = NULL,
  facet_grid = NULL,
  xlab = "",
  ylab = "",
  title = "",
  guideTitleSize = "",
  alpha = 0.5,
  maxBubbleSize = 6,
  useColourGradient = FALSE,
  colourPalette = wtsUtilities::createColorPalette("jet", 100, alpha = alpha),
  guideTitleColour = "",
  ggtheme = ggplot2::theme_gray(),
  showPlot = FALSE
)
}
\arguments{
\item{mdfr}{- melted dataframe}

\item{x}{- column name for x-axis values}

\item{y}{- column name for y-axis values}

\item{value.var}{- column name for values to aggregate (value.var in cast)/plot as circles}

\item{agg.formula}{- aggregation formula (left-hand side of cast formula)}

\item{agg.function}{- aggregation function (fun.aggregate in cast)}

\item{...}{- further arguments passed to aggregating function}

\item{colour}{- column name to which colour aesthetic is mapped}

\item{facet_grid}{- faceting formula}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{title}{- plot title}

\item{guideTitleSize}{- title for size guide}

\item{alpha}{- transparency level}

\item{maxBubbleSize}{- max bubble size}

\item{useColourGradient}{- flag (T/F) to use a color gradient for bubble color}

\item{colourPalette}{- colour palette for colour gradient (default=\code{wtsUtilities::createColorPalette('jet',100,alpha=alpha)})}

\item{guideTitleColour}{- title for colour guide}

\item{ggtheme}{- ggplot2 theme for plot (default=\code{ggplot2::theme_gray()})}

\item{showPlot}{- flag to show plot immediately}
}
\value{
ggplot2 object
}
\description{
Function to plot a melted dataframe using a bubble plot format.
}
\details{
uses \code{reshape2} package.
}
