% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDFR.XY.R
\name{plotMDFR.XY}
\alias{plotMDFR.XY}
\title{Plot melted dataframe using an xy-plot format.}
\usage{
plotMDFR.XY(
  mdfr,
  x = NULL,
  value.var = "val",
  agg.formula = NULL,
  agg.function = wtsUtilities::Sum,
  colour = NULL,
  fill = NULL,
  linetype = NULL,
  shape = NULL,
  facet_grid = NULL,
  facet_wrap = NULL,
  nrow = NULL,
  ncol = NULL,
  dir = "h",
  scales = "fixed",
  xlim = NULL,
  ylim = NULL,
  xlab = "",
  ylab = "",
  units = "",
  lnscale = FALSE,
  title = "",
  guideTitleColour = NULL,
  guideTitleFill = NULL,
  guideTitleLineType = NULL,
  guideTitleShape = NULL,
  plotPoints = TRUE,
  points = list(alpha = 1, size = 1, dodge = 0),
  plotLines = TRUE,
  plotABline = FALSE,
  abline = list(intercept = 0, slope = 1, colour = "black", linetype = 3, size = 1,
    alpha = 0.8),
  plotViolins = FALSE,
  violins = list(scale = "count", stat = "ydensity", dodge = 0, alpha = 1),
  ggtheme = ggplot2::theme_gray(),
  showPlot = FALSE
)
}
\arguments{
\item{mdfr}{- melted dataframe}

\item{x}{- column name for x-axis values}

\item{value.var}{- column name for values to aggregate (value.var in cast)/plot on y-axis}

\item{agg.formula}{- aggregation formula (left-hand side of cast formula)}

\item{agg.function}{- aggregation function (fun.aggregate in cast, default=wtsUtilities::Sum)}

\item{colour}{- column name to which colour aesthetic is mapped}

\item{fill}{- column nameto which fill aesthetic is mapped}

\item{linetype}{- column name to which linetype aesthetic is mapped}

\item{shape}{-  column name to which shape aesthetic is mapped}

\item{facet_grid}{- faceting formula for facet_grid'ing}

\item{facet_wrap}{- faceting formula for facet_wrap'ing}

\item{nrow}{- number of rows when using facet_wrap}

\item{ncol}{- number of columns when using facet_wrap}

\item{dir}{- fastest direction ('h','v') for faceting}

\item{scales}{- scaling flag when faceting}

\item{xlim}{- x axis limits}

\item{ylim}{- y axis limits}

\item{xlab}{- x axis label}

\item{ylab}{- y axis label}

\item{units}{- combined with y axis label}

\item{lnscale}{- flag to plot on ln-scale}

\item{title}{- plot title}

\item{guideTitleColour}{- title for colour guide}

\item{guideTitleFill}{- title for fill guide}

\item{guideTitleLineType}{- title for linetype guide}

\item{guideTitleShape}{- title for shape guide}

\item{plotPoints}{- flag to plot points}

\item{points}{- list with elements alpha=1.0,size=1.0,dodge=0.0}

\item{plotLines}{- flag to plot lines}

\item{plotABline}{- flag to plot a straight line}

\item{abline}{- list w/ components intercept, slope, colour, size, linetype, alpha describing line to plots}

\item{plotViolins}{- flag to plot violins}

\item{violins}{- list with elements scale="count",stat="ydensity",dodge=0.0,alpha=1.0 (see \code{ggplot2::geom_violin} for details)}

\item{ggtheme}{- ggplot2 theme for plot (default=\code{ggplot2::theme_gray()})}

\item{showPlot}{- flag to show plot immediately}
}
\value{
ggplot2 object
}
\description{
Function to plot a melted dataframe using an xy plot format.
}
\details{
uses \code{reshape2} package.
}
