% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRStanMCMC.R
\name{plotRStanMCMC}
\alias{plotRStanMCMC}
\title{Make diagnostic plots for a parameter in an [rstan](https://mc-stan.org/) mcmc simulation}
\usage{
plotRStanMCMC(mcmc, var, label = var)
}
\arguments{
\item{mcmc}{- [rstan](https://mc-stan.org/) stanfit object from running [rstan::sampling()] or [tmbstan::tmbstan()]}

\item{var}{- name of variable to plot}

\item{label}{- label to use for variable (default=\code{var})}
}
\value{
a ggplot2::gg object (or NULL if the [rstan](https://mc-stan.org/) package is not installed.)
}
\description{
Function to make diagnostic plots for a parameter in an [rstan](https://mc-stan.org/) mcmc simulation.
}
\details{
Requires the [rstan](https://mc-stan.org/) and [cowplot] packages. Combines plots from running [rstan::stan_trace()], [rstan::stan_ac()] (both ACF and PACF plots),
[rstan::stan_hist()], and [rstan::density()] for the model parameter identified by \code{var} Plots 
are combined using [cowplot::plot_grid()]. The \code{label} input can be specified to provide a 
more informative name for the plotted parameter.
}
