% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal_Hist2D.R
\name{print.ggGTbl}
\alias{print.ggGTbl}
\title{Print a \code{ggGTbl} object}
\usage{
\method{print}{ggGTbl}(x, newpage = grDevices::dev.interactive(), ...)
}
\arguments{
\item{x}{- ggGTbl object.}

\item{newpage}{- flag to draw a new (empty) page before the ggTbl object is drawn?}

\item{...}{ignored}
}
\description{
\code{ggGTbl} objects are created from \code{ggMarginal_XX} functions. 
This is the S3 generic print method to print the ggGTbl result of the 
function.
}
\note{
This function is based on the [ggExtra::ggMarginal()] 
code developed by Dean Attali ("daattali@gmail.com").
}
\seealso{
[ggMarginal_Hist2D]
}
\keyword{internal}
