% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggMarginal_Hist2D.R
\name{rbindGrobs}
\alias{rbindGrobs}
\title{Add a grob to the top of a gtable}
\usage{
rbindGrobs(topGrob, gtable, l, r)
}
\arguments{
\item{topGrob}{- grob to add to gtable as top row}

\item{gtable}{- gtable to be added to}

\item{l}{- left size}

\item{r}{- right size}
}
\description{
Function to add a grob to the top of a gtable.
}
\details{
Copied from \code{ggExtra}. Created by Dean Attali ("daattali@gmail.com").
}
\keyword{internal}
